# src/main/resources/public/swagger/swagger.yaml
openapi: '3.0.3'
info:
  title: baragym-backend
  version: 0.0.1
  description: Backend da aplicação BaraGym.
servers:
  - url: 'https://baragym-backend.herokuapp.com/api/v1'
  - url: 'http://localhost:8080/api/v1'
...
paths:
...
  /atividades:
    post:
      tags:
        - atividade
      summary: Cria uma atividade
      description: Cria uma atividade a partir do nome, descricao e valor por dia.
      security:
        - bearerAuthToken: [ ]
      requestBody:
        description: Corpo da requisição de criação de atividade
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAtividadeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtividadeResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    get:
      tags:
        - atividade
      summary: Obtém todas as atividades
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: page
          in: query
          description: |
            Número de página solicitado. <br/>
            **Atenção: os números de página inciam em `0`.**
          required: false
          schema:
            type: integer
            example: 0
            default: 0
        - name: size
          in: query
          description: Tamanho máximo da página de dados retornada.
          required: false
          schema:
            type: integer
            example: 20
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtividadePageResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
...
  /atividades/{atividadeId}:
    get:
      tags:
        - atividade
      summary: |
        Obtem uma atividade
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: atividadeId
          in: path
          description: Identificador único da atividade
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtividadeResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    put:
      tags:
        - atividade
      summary: |
        Atualiza uma atividade
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: atividadeId
          in: path
          description: Identificador único da atividade
          required: true
          schema:
            type: string
      requestBody:
        description: Corpo da requisição de atualização da atividade.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAtividadeRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    delete:
      tags:
        - atividade
      summary: |
        Deleta uma atividade
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: atividadeId
          in: path
          description: Identificador único da atividade
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno

  ######################
  ###### ALOCACAO ######
  ######################
  /alocacoes:
    post:
      tags:
        - alocacao
      summary: Cria uma alocacao
      description: |
        Cria uma nova alocacao. Baseado na turma, atividade e dia da semana.
      security:
        - bearerAuthToken: [ ]
      requestBody:
        description: Corpo da requisição de criação de alocacao.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlocacaoRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    get:
      tags:
        - alocacao
      summary: Obtém todas as alocacoes
      description: Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: page
          in: query
          description: |
            Número de página solicitado. <br/>
            **Atenção: os números de página inciam em `0`.**
          required: false
          schema:
            type: integer
            example: 0
            default: 0
        - name: size
          in: query
          description: Tamanho máximo da página de dados retornada.
          required: false
          schema:
            type: integer
            example: 20
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlocacaoPageResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
  /atividades/{atividadeId}/alocacoes:
    get:
      tags:
        - alocacao
      summary: Obtém as alocações de determinada atividade
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: atividadeId
          in: path
          description: Identificador da atividade
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: |
            Número de página solicitado. <br/>
            **Atenção: os números de página inciam em `0`.**
          required: false
          schema:
            type: integer
            example: 0
            default: 0
        - name: size
          in: query
          description: Tamanho máximo da página de dados retornada.
          required: false
          schema:
            type: integer
            example: 20
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlocacaoPageResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
  /usuarios/{matricula}/atividades/{usuarioAtividadeId}/alocacoes:
    get:
      tags:
        - alocacao
      summary: |
        Obtem, a partir das atividades cadastradas do usuário, suas alocacoes e se possui agendamentos
      description: |
        Apenas o próprio usuário tem acesso ou um administrador (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: matricula
          in: path
          description: Matrícula do usuário
          required: true
          schema:
            type: string
        - name: usuarioAtividadeId
          in: path
          description: Identificador único do cadastro do usuário em uma determinada atividade
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: |
            Número de página solicitado. <br/>
            **Atenção: os números de página inciam em `0`.**
          required: false
          schema:
            type: integer
            example: 0
            default: 0
        - name: size
          in: query
          description: Tamanho máximo da página de dados retornada.
          required: false
          schema:
            type: integer
            example: 20
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioAlocacaoAgendamentoPageResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
  /alocacoes/{alocacaoId}:
    get:
      tags:
        - alocacao
      summary: |
        Obtém uma alocacao
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: alocacaoId
          in: path
          description: Identificador único da alocacao
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlocacaoResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    put:
      tags:
        - alocacao
      summary: |
        Atualiza uma alocacao
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: alocacaoId
          in: path
          description: Identificador único da alocacao
          required: true
          schema:
            type: string
      requestBody:
        description: Corpo da requisição de atualização da alocacao.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlocacaoRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    delete:
      tags:
        - alocacao
      summary: |
        Deleta uma alocacao
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: alocacaoId
          in: path
          description: Identificador único da alocacao
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
components:
  securitySchemes:
    bearerAuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acesso à aplicação.
  schemas:
...
    SortResponse:
      type: object
      properties:
        empty:
          type: boolean
          example: true
          description: Verifica se a lista está vazia ou não.
        sorted:
          type: boolean
          example: false
          description: Verifica se a lista está ordenada ou não.
        unsorted:
          type: boolean
          example: true
          description: Verifica se a lista está fora de ordenação ou não.
...
    UsuarioAlocacaoAgendamentoResponse:
      type: object
      properties:
        agendamentoId:
          type: number
          description: Identificador único do agendamento
          example: 1
        alocacaoId:
          type: number
          description: Identificador único da alocacao
          example: 3
        nomeAtividade:
          type: string
          description: Nome da atividade da alocacao
          example: 'Boxe'
        nomeDiaSemana:
          type: string
          description: Dia da semana da alocacao
          example: 'Terça-feira'
        nomeTurma:
          type: string
          description: Nome da turma/sala da alocacao
          example: 'Sala 1'
        horaInicio:
          type: string
          description: Horário de início da alocacao
          example: '10:00'
        horaFim:
          type: string
          description: Horário de término da alocacao
          example: '15:00'
        estaAgendado:
          type: boolean
          description: |
            Verifica se esta alocacao possui um agendamento prévio, 
            caso seja verdade, o campo `agendamentoId` também virá preenchido.
          example: true
    UsuarioAlocacaoAgendamentoPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UsuarioAlocacaoAgendamentoResponse'
        pageable:
          type: object
        last:
          type: boolean
          example: false
          description: Verifica se é a última página.
        totalElements:
          type: number
          example: 5
          description: Número de elementos totais.
        totalPages:
          type: number
          example: 5
          description: Número de páginas totais.
        size:
          type: number
          example: 1
          description: Número de elementos por página.
        number:
          type: number
          example: 0
          description: Número de da página.
        sort:
          $ref: '#/components/schemas/SortResponse'
        first:
          type: boolean
          example: true
          description: Verifica se é a primeira página.
        numberOfElements:
          type: number
          example: 5
          description: Número de elementos totais.
        empty:
          type: boolean
          example: false
          description: Verifica se a lista está vazia ou não.
...
    CreateAtividadeRequest:
      type: object
      properties:
        nome:
          type: string
          description: Nome da atividade
          example: 'Boxe'
        descricao:
          type: string
          description: Descricao da atividade
          example: 'Muito bom para quem quer queimar gordura...'
        valorDia:
          type: number
          description: Valor por dia da atividade
          example: 50.99
    AtividadeResponse:
      type: object
      properties:
        id:
          type: number
          description: identificador único da atividade
          example: 12
        nome:
          type: string
          description: Nome da atividade
          example: 'Boxe'
        descricao:
          type: string
          description: Descricao da atividade
          example: 'Muito bom para quem quer queimar gordura...'
        valorDia:
          type: number
          description: Valor por dia da atividade
          example: 50.99
    AtividadePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AtividadeResponse'
        pageable:
          type: object
        last:
          type: boolean
          example: false
          description: Verifica se é a última página.
        totalElements:
          type: number
          example: 5
          description: Número de elementos totais.
        totalPages:
          type: number
          example: 5
          description: Número de páginas totais.
        size:
          type: number
          example: 1
          description: Número de elementos por página.
        number:
          type: number
          example: 0
          description: Número de da página.
        sort:
          $ref: '#/components/schemas/SortResponse'
        first:
          type: boolean
          example: true
          description: Verifica se é a primeira página.
        numberOfElements:
          type: number
          example: 5
          description: Número de elementos totais.
        empty:
          type: boolean
          example: false
          description: Verifica se a lista está vazia ou não.
    CreateAlocacaoRequest:
      type: object
      properties:
        turmaId:
          type: number
          description: Identificador único da turma
          example: 12
        atividadeId:
          type: number
          description: Identificador único da atividade
          example: 3
        diaSemanaId:
          type: number
          description: Identificador único do dia da semana
          example: 4
        horaInicio:
          type: string
          description: Horário de início da alocação
          example: '15:00'
        horaFim:
          type: string
          description: Horário de fim da alocação
          example: '17:00'
    AlocacaoResponse:
      type: object
      properties:
        id:
          type: number
          description: identificador único da alocacao
          example: 12
        atividade:
          description: nome e id da atividade da alocacao
          $ref: '#/components/schemas/IdAndNomeResponse'
        diaSemana:
          description: nome e id do dia da semana da alocacao
          $ref: '#/components/schemas/IdAndNomeResponse'
        turma:
          description: nome e id da turma da alocacao
          $ref: '#/components/schemas/IdAndNomeResponse'
        horaInicio:
          type: string
          description: Horário de início da alocacao
          example: '15:00'
        horaFim:
          type: string
          description: Horário de término da alocacao
          example: '17:00'
    AlocacaoPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AlocacaoResponse'
        pageable:
          type: object
        last:
          type: boolean
          example: false
          description: Verifica se é a última página.
        totalElements:
          type: number
          example: 5
          description: Número de elementos totais.
        totalPages:
          type: number
          example: 5
          description: Número de páginas totais.
        size:
          type: number
          example: 1
          description: Número de elementos por página.
        number:
          type: number
          example: 0
          description: Número de da página.
        sort:
          $ref: '#/components/schemas/SortResponse'
        first:
          type: boolean
          example: true
          description: Verifica se é a primeira página.
        numberOfElements:
          type: number
          example: 5
          description: Número de elementos totais.
        empty:
          type: boolean
          example: false
          description: Verifica se a lista está vazia ou não.
    IdAndNomeResponse:
      type: object
      properties:
        id:
          type: number
          example: 1
          description: Identificador único
        nome:
          type: string
          example: nomeDominio
          description: Nome do dominio